pyproject.toml
#[build-system]
#requires = ["setuptools", "wheel"]
#build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["moonwatcher*"]
exclude = ["docs*", "tests*"]

[tool.poetry]
name = "moonwatcher"
version = "0.1.0-alpha1"
description = ""
authors = ["Leon Steffen <hello@moonwatcher.ai>"]
readme = "README.md"
license = "Apache-2.0"
homepage = "https://www.moonwatcher.ai"
repository = "https://github.com/moonwatcher-ai/moonwatcher"

[tool.poetry.urls]
"Issues" = "https://github.com/moonwatcher-ai/moonwatcher/issues"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2.2"
torchvision = "^0.17.2"
requests = "^2.31.0"
numpy = "^1.26.4"
pytz = "^2024.1"
tqdm = "^4.66.2"
opencv-python = "^4.9.0.80"
torchmetrics = "^1.3.2"
scikit-learn = "^1.4.1.post1"
transformers = "^4.39.3"
timm = "^0.9.16"
setuptools = "^69.2.0"
pandas = "^2.2.1"
python-dotenv = "^1.0.1"
pycocotools = "^2.0.7"
gdown = "^5.1.0"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
